improve E2E test helper folder structure
